fastapi/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["kathmandu777"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
SQLAlchemy = "^1.4.27"
aiomysql = "^0.0.21"
psycopg2-binary = "^2.9.2"
alembic = {extras = ["tz"], version = "^1.7.5"}
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytz = "^2021.3"
python-jose = {version = "3.2.0", extras = ["cryptography"]}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
